Updated specification
Introduced the concept of the eson_pair as a superset of the JSON_pair. Introduced the identifier as the duplicate key restriction means I can't use de\
f more than once to make variables. Clarified the definition of a record and included attributes. Records are nested eson documents while attributes are members of the eson document that SHALL appear in the JSON equivalent document.