Created Language module to hold eson language
Drafted a specification for how the grammar/language can be expressed as a data structure. An initial draft of the data structure is also given.
Moved ebnf language rules to Language module. I intend to use Language as a singleton from which I can produce the grammar's production rules and also create derivative languages similar to the nanopass framework.