Cleaned up Language#initial specification
Listed rules with non-terminals to the top to make it easier to define the production rules as methods Slightly changed the Structs to use more correct EBNF terminology, also replaced =, 'equal' signs with :=, 'defined as' in production rules. I notice some recursive definitions in the production rules but I'm not sure if those are an issue in terms of defining the non-terminals. I'll find out when I write the methods to create the rules.