Reorganized Language module
The initial formal language for the eson compiler has been reduced to tokens only. This is to enable early passes before the Parser stage to remove ambiguity for the recursive descent algorithm and to do early error checking and analysis.
Will try to allow each pass to operate on a different language that can be dynamically generated as hand crafting E0, the initial language was tedious. The work I've done with regexp's looks promising in this regard and hopefully all additional formal languages will be built off E0.
The initial formal Language is that used by the tokenizer to create the token sequence and contains only terminal production rules. Refactored Tokenizer module to use the Language module for some of it's pattern matching operations for eson Tokens. This made it necessary to add functionality to the Language module including:
- Rule operations for invoking pattern matching on Terminals using methods
- Language operation for creating an alternation rule from terminal rules
There is much more than can be done to make Tokenizer more dependent on the initial language produced by the Language module. This coupling of the modules is to enhance the cohesion between the data structures representing the language and the operations in the Tokenizer.