Refactored Tokenizer
Tokenizer uses more of the primitives in Language module to accomplish tokenization. However there is still a great deal of repetition involved and I see an opportunity to automate more of this based on the Language rules but I haven't figured out how to do that yet.
Also added another NonTerminal production rule to e0, word_forms in preparation for tokenizing strings with variable_identifiers.