Added active_node method to AST
Added method to determine the active_node in the tree. It currently works only for the case of the newly created tree in which the active node is the root node. This also introduces the idea of the open and closed tree node. Both can be checked using the open? and closed? methods respectively. All newly created tree nodes are given an open status of true via a struct field.