Insert_node adds only valid tokens
Insert node will check if token is a valid next node before inserting to the active node. It will throw a TreeInsertionError if the token is not valid. A token is valid if when added the ordered list of child nodes still holds as a valid sequence for the node's rule. So far this works only for a node with a concatenation rule where the token is the first child node to be added to the active node.