Converted Rule to a pure class
Needing to do intialization tasks with the Rule class for computing first_sets, I could not continue to use a Struct for representing a Rule. I tried using modules and prepend to hook into the initialization of the Struct but there were still alot of issues with that implementation. Namely, how does one access the properties of the object before it is created.