Computes first sets for each rule
First sets are used in the recursive descent algorithm used to build the parse tree, this is the set of tokens that can legally appear at the start of a sequence for a production rule. Upon initialization each production rule is given a first_set, a first set will be partially complete if it's definition contains recursive terms or terms that descend from recursive terms. When a formal language is built from a set of rules all partially complete rules are completed.