Refactored recursive and partial concept
Recursive and partial concepts overlapped in the implementation so they have been combined into a redefined partial concept. All terminals have a partial status of false, while a nonterminal will have a true partial status if any of it's terms are undefined or a descendant of an undefined term. An undefined term is any term that is not present in the RuleSeq at the time at which the Rule is created.
As occured previously, once a language is built from the RuleSeq, the partial status is set to false for all Rules and all first sets are computed. The recursive term has also been removed from the NonTerminal definition which means there is currently no difference between the fields in the Terminal and the NonTerminal.