Implemented error path for terminal parsing
Produces a simple error message comparing the expected token name to the actual token name. I would prefer if this could display lexemes instead but the Language module does not have any knowledge of the lexemes. This is strange, since the Tokenizer module magically knows the lexemes of the rules contained in the FormalLanguage module. This relationship needs to be inverted, lexemes should be present in terminal rules and rules should define a method to produce tokens.