Follow sets computed for rules
When language is built, follow_sets are generated for each rule. The follow_set is initialized as an empty_array and the top_rule of the language has :eof added to denote the end of file symbol. The implementation has a functional style and does not use a partial status like the first_set computation.