Emergency commit
There is a parsing bug that is affecting the array and program rules, which has not yet been resolved. This commit contains refactoring and edits that have taken place so far to remove the bug. The bug still exists so the test for the SyntaxPass module is failing, checking in changes now because I need to work from another machine.