Tokenizer exception added
Based on the specification for the `tokenize_program` the program can detect when a program has not been tokenized correctly. Previously this check was done in the test suite, which meant that the method had to produce an additional return which was just used for testing for this exception. This also meant that in regular use this error condition would not be picked up when it happened. This has been rectified by including the TokenizationIncomplete exception which is raised if the error condition ever occurs. Because this error condition should never occur the error message instructs the user to file a bug report.