Added module to track receiver dependencies in Mixins
Implemented a module called Respondent which allows one to document the receiver dependencies of a module and check if they are satisfied. Using :uses in a module I can supply all the methods that the receiver (module or class that is including or extending) must provide for the correct operation of the module. I can validate this in the definition of the receiver by calling the .validate method on the module. If a method isn't found it will throw a MethodMissing exception.