Set line_feed s-attribute in tokenizer
Implemented the logic to set line_feed attribute during tokenization. All tokens are now assigned this s-attribute as it is not restricted to the set to which it was initially assigned.
When a line_feed attribute is set to true a line_feed must follow the lexeme of the token.