Implemented ParseTree#shift_root
Method to shift the root of a tree downwards only when that root has a single child. This method may change the height of the tree but does not yet try to compute the correct hieght. This method deletes the root node of the ParseTree or ParseTree::Tree and will be used to remove alternation, option and repetition type nodes from the parse tree.