Alternation rules removed from AST
AST does not include alternation rules as root nodes. With the sample program used in the ast tests, the height of the tree was halved by the removal of alternation nodes.
An attempt was made to preserve the alternation node as an attribute but this couldn't fit into the rules of the attribute grammar concept.