Syntax representation for ag rules
Using EBNF module to provide syntax definitions for the attribute grammar production rules. This is intended to to better represent the structure of the Abstract Syntax Tree which uses attribute grammar rules. The syntax definition would providing a definition of the node structure, which can then be used to dynamically apply the transformations in the Dote::DoteGrammars::AST module. This approach also makes it easier to maintain the AST because the structure of the AST will be available in a declarative form.